+++ lib/adhearsion/generators/app/templates/lib/routes.rb
@@ -1,0 +1,10 @@
+# encoding: utf-8
+
+Adhearsion.router do
+
+  #
+  # Specify your call routes, directing calls with particular attributes to a controller
+  #
+
+  route 'default', SimonGame
+end